requirements.txt
requests==2.25.0
Pillow==8.0.1
instabot==0.117.0
python-dotenv==0.15.0

